requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.7
Werkzeug==3.0.3
python-dotenv==1.0.1
WTForms==3.1.2
openpyxl==3.1.5
reportlab==4.2.5
requests==2.32.3
woocommerce==3.0.0
email-validator==2.2.0
Pillow==10.4.0
gunicorn==22.0.0
